Add the trustable trait.
Also bump version to 0.4.6-SNAPSHOT.
