+++ core/src/test/scala/com.soundcloud.sketchy/SpecHelper.scala
@@ -242,8 +242,6 @@ trait SpecHelper {
       def senderId = Some(_userId)
       def content = _body
 
-      var trusted: Option[Boolean] = None
-      var interaction: Option[Boolean] = None
       val public: Option[Boolean] = None
 
       def noSpamCheck = admMsgOrTrusted
